+++ pojo/src/main/java/com/stanzaliving/item_master/dtos/UpdatedBoqItemRequest.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -15,5 +16,5 @@ public class UpdatedBoqItemRequest {
     private String itemUuid;
     private boolean status;
     private Date lastUpdatedAt;
-    private Double rate;
+    private BigDecimal rate;
 }
